Rework how simulation progress is displayed in the UI
- Progress bar and stop button are no longer shown in text area. They're
  shown in the status panel just above the textview widget
- Simulation status (ie "X of Y completed") is shown in a label to the
  right of the progress bar, above the text area
- Errors in the simulation are displayed in the text area as they occur

Other changes not really relevant to users:

- Fixed the "collection has been modified" error which was being thrown
  when updating simulation progress. (This was being handled internally
  anyway, so was only visible when running with a debugger.)
- Fixed the progress bar appearance in gtk3 builds. I've set it to a
  fixed 34px height. It would be better to have it expand to the same
  size as the stop button.

+++ Tests/UnitTests/Core/Run/GenerateApsimXFilesTests.cs
@@ -65,7 +65,7 @@
             Directory.CreateDirectory(path);
 
             // Create a list of progress ints.
-            var progress = new List<int>();
+            var progress = new List<double>();
 
             // Create a runner for our folder.
             Runner runner = new Runner(folder);
